Additional refactoring

+++ middle_man.rb
@@ -3,7 +3,6 @@ require './user_interface'
 require './database'
 
 class MiddleMan
-
   attr_reader :id, :action, :task, :database, :user_interface
 
   def initialize(args)
@@ -17,8 +16,11 @@ class MiddleMan
 
   def execute!
     if self.respond_to?(action.to_sym)
-      self.send(action.to_sym)
-      database.save
+      if self.send(action.to_sym)
+        database.save
+      else
+        user_interface.invalid_id
+      end 
     else
       user_interface.non_action
     end
@@ -31,6 +33,7 @@ class MiddleMan
   def add
     database.add_item(ListItem.new(default_task_args))
     user_interface.confirm_add(task)
+    true
   end
 
   def get_list
@@ -43,13 +46,19 @@ class MiddleMan
  
   def delete
     item = database.get_item(id)
+    return false if item.nil?
+
     database.delete(item.id)
     user_interface.confirm_delete(item.task)
+    true
   end
 
   def complete
     item = database.get_item(id)
+    return false if item.nil?
+
     database.complete_item(item.id)
     user_interface.confirm_complete(item.task)
+    true
   end
 end
